backend/requirements.txt
django
djangorestframework
pygments
Pillow